// src/app/movies/movies.service.interface.ts
export type Movie = {
    budget: number;
    genres: { id: number; name: string }[];
    homepage: string;
    id: number;
    keywords: { id: number; name: string }[];
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    production_companies: { id: number; name: string }[];
    production_countries: { iso_3166_1: string; name: string }[];
    release_date: string;
    revenue: number;
    runtime: number;
    spoken_languages: { iso_639_1: string; name: string }[];
    status: string;
    tagline: string;
    title: string;
    vote_average: number;
    vote_count: number;
};